patch from christopherkobayashi : << Hello,
First, thank you for putting together the software to effectively replace a floppy drive. I have replaced all of the floppies in my retro equipment with HxCs, and I am very happy with them.

I have found a bug, however, in the way that the TI-99/4A V9T9 format writer handles a certain type of floppy format. It has become common in the TI community to use modified firmware with stock TI disk controllers that allow 80 tracks per side, rather than the stock 40, but still at single density and using FM encoding. This is because third-party controllers are extremely difficult to find and rather expensive.

The V9T9 format writer chooses the disk image format based on image size, and in this case, it chooses the wrong one (40-track, double-density, MFM encoding). This results in an inability to convert .hfe back to V9T9.

Here is a patch that tries to read at 80tps DSSD FM if a) an initial read of track 0/sector 0/side 0 fails and b) the flag "fm_fallback" is set. I hope that this patch is acceptable.

Thanks,

-- Chris>>

git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code@1389 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69